chore(team): split mailbox manager tests
