+++ src/team/manager/tests.rs
@@ -6,6 +6,7 @@ mod archive_cases;
 mod cancel_cases;
 mod channel_cases;
 mod linked_run_cases;
+mod mailbox_basic_cases;
 mod reconcile_cases;
 mod run_admin_cases;
 mod run_creation_cases;
@@ -2236,1335 +2237,6 @@ async fn list_steps_returns_sorted_steps_for_a_run() {
             .collect::<Vec<_>>(),
         vec!["a-step", "z-step"]
     );
-}
-
-#[tokio::test]
-async fn actor_messages_support_inbox_and_ack_flow() {
-    let db = setup_test_db().await;
-    let manager = TeamManager::new(db.clone());
-
-    let team = manager
-        .create_team(TeamDefinitionConfig {
-            name: "actor-message-team".to_string(),
-            description: Some("team for actor message flow".to_string()),
-            spec: json!({
-                "entrypoint":"planner",
-                "members":[{"member_id":"planner"},{"member_id":"reviewer"}]
-            }),
-        })
-        .await
-        .expect("create team");
-    let run = manager
-        .create_run(&team.id, Some("ctx-msg"), json!({"payload":"start"}))
-        .await
-        .expect("create run");
-
-    let sent = manager
-        .send_actor_message(SendActorMessageInput {
-            run_id: &run.id,
-            from_actor_id: "planner",
-            from_peer_id: ACTOR_MAIN_PEER_ID,
-            to_actor_id: "reviewer",
-            to_peer_id: ACTOR_MAIN_PEER_ID,
-            channel: "coordination",
-            transport: TeamActorMessageTransport::Local,
-            route: None,
-            payload: json!({"text":"please review"}),
-            idempotency_key: None,
-            message_kind: None,
-        })
-        .await
-        .expect("send message");
-    let expected_payload = json!({
-        "text":"please review",
-        "source_kind":"agent",
-        "source_surface":"mailbox",
-        "requires_user_visible_reply":false
-    });
-    assert_eq!(sent.status, TeamActorMessageStatus::Pending);
-    assert_eq!(sent.transport, TeamActorMessageTransport::Local);
-    assert_eq!(sent.payload, expected_payload);
-    assert_eq!(sent.from_actor_kind, ActorIdentityKind::Agent);
-    assert_eq!(sent.to_actor_kind, ActorIdentityKind::Agent);
-
-    let human_sent = manager
-        .send_actor_message(SendActorMessageInput {
-            run_id: &run.id,
-            from_actor_id: "user:alice",
-            from_peer_id: ACTOR_MAIN_PEER_ID,
-            to_actor_id: "reviewer",
-            to_peer_id: ACTOR_MAIN_PEER_ID,
-            channel: "coordination",
-            transport: TeamActorMessageTransport::Local,
-            route: None,
-            payload: json!({"text":"human request"}),
-            idempotency_key: None,
-            message_kind: None,
-        })
-        .await
-        .expect("send human message");
-    assert_eq!(human_sent.from_actor_kind, ActorIdentityKind::Human);
-    assert_eq!(human_sent.to_actor_kind, ActorIdentityKind::Agent);
-
-    let pending_inbox = manager
-        .list_actor_inbox(&run.id, "reviewer", 100, None, false)
-        .await
-        .expect("list pending inbox");
-    assert_eq!(pending_inbox.len(), 2);
-    let pending_ids = pending_inbox
-        .iter()
-        .map(|message| message.message_id)
-        .collect::<Vec<_>>();
-    assert!(pending_ids.contains(&sent.message_id));
-    assert!(pending_ids.contains(&human_sent.message_id));
-    let human_inbox = pending_inbox
-        .iter()
-        .find(|message| message.message_id == human_sent.message_id)
-        .expect("human message in inbox");
-    assert_eq!(human_inbox.from_actor_kind, ActorIdentityKind::Human);
-    assert_eq!(human_inbox.to_actor_kind, ActorIdentityKind::Agent);
-
-    let delivered = manager
-        .ack_actor_message(&run.id, "reviewer", sent.message_id)
-        .await
-        .expect("ack message");
-    assert!(delivered.status_changed);
-    assert_eq!(delivered.message.status, TeamActorMessageStatus::Delivered);
-    assert!(delivered.message.delivered_at.is_some());
-
-    let pending_after_ack = manager
-        .list_actor_inbox(&run.id, "reviewer", 100, None, false)
-        .await
-        .expect("list pending after ack");
-    assert_eq!(pending_after_ack.len(), 1);
-    assert_eq!(pending_after_ack[0].message_id, human_sent.message_id);
-    assert_eq!(
-        pending_after_ack[0].from_actor_kind,
-        ActorIdentityKind::Human
-    );
-
-    let inbox_with_delivered = manager
-        .list_actor_inbox(&run.id, "reviewer", 100, None, true)
-        .await
-        .expect("list inbox with delivered");
-    assert_eq!(inbox_with_delivered.len(), 2);
-    let delivered_message = inbox_with_delivered
-        .iter()
-        .find(|message| message.message_id == sent.message_id)
-        .expect("delivered message exists");
-    assert_eq!(delivered_message.status, TeamActorMessageStatus::Delivered);
-
-    let events = manager
-        .list_run_events(&run.id, 100, None)
-        .await
-        .expect("list run events");
-    let event_types: Vec<&str> = events
-        .iter()
-        .map(|event| event.event_type.as_str())
-        .collect();
-    assert_eq!(
-        event_types,
-        vec![
-            "run_submitted",
-            "actor_message_sent",
-            "actor_message_sent",
-            "actor_message_delivered"
-        ]
-    );
-}
-
-#[tokio::test]
-async fn summarize_open_reply_obligations_prefers_lightweight_snapshot_loader() {
-    let db = setup_test_db().await;
-    let manager = TeamManager::new(db);
-
-    let team = manager
-        .create_team(TeamDefinitionConfig {
-            name: "reply-obligation-team".to_string(),
-            description: Some("team for reply obligation summary".to_string()),
-            spec: json!({
-                "entrypoint":"planner",
-                "members":[{"member_id":"planner"},{"member_id":"reviewer"}]
-            }),
-        })
-        .await
-        .expect("create team");
-    let run = manager
-        .create_run(
-            &team.id,
-            Some("ctx-reply-obligation"),
-            json!({"payload":"start"}),
-        )
-        .await
-        .expect("create run");
-
-    manager
-        .send_actor_message(SendActorMessageInput {
-            run_id: &run.id,
-            from_actor_id: "user:alice",
-            from_peer_id: ACTOR_MAIN_PEER_ID,
-            to_actor_id: "reviewer",
-            to_peer_id: ACTOR_MAIN_PEER_ID,
-            channel: "coordination",
-            transport: TeamActorMessageTransport::Local,
-            route: None,
-            payload: json!({
-                "type":"chat_message",
-                "text":"Need update",
-                "requires_user_visible_reply":true
-            }),
-            idempotency_key: None,
-            message_kind: None,
-        })
-        .await
-        .expect("send inbound obligation");
-    manager
-        .send_actor_message(SendActorMessageInput {
-            run_id: &run.id,
-            from_actor_id: "reviewer",
-            from_peer_id: ACTOR_MAIN_PEER_ID,
-            to_actor_id: "user:alice",
-            to_peer_id: ACTOR_MAIN_PEER_ID,
-            channel: "coordination",
-            transport: TeamActorMessageTransport::Local,
-            route: None,
-            payload: Value::String(
-                r#"{"type":"chat_message","text":"Here is the update","correlation_id":"corr-summary"}"#
-                    .to_string(),
-            ),
-            idempotency_key: None,
-            message_kind: None,
-        })
-        .await
-        .expect("send visible reply");
-
-    let summary = manager
-        .summarize_open_reply_obligations(&run.id)
-        .await
-        .expect("summarize reply obligations");
-
-    assert_eq!(summary.open_total, 0);
-    assert!(summary.open_by_actor.is_empty());
-    assert!(summary.open_items.is_empty());
-}
-
-#[tokio::test]
-async fn actor_messages_detect_pending_payload_type_by_actor_inbox() {
-    let db = setup_test_db().await;
-    let manager = TeamManager::new(db);
-
-    let team = manager
-        .create_team(TeamDefinitionConfig {
-            name: "actor-message-payload-type-team".to_string(),
-            description: Some("team for payload type pending lookup".to_string()),
-            spec: json!({
-                "entrypoint":"planner",
-                "members":[{"member_id":"planner"},{"member_id":"reviewer"}]
-            }),
-        })
-        .await
-        .expect("create team");
-    let run = manager
-        .create_run(&team.id, Some("ctx-msg-type"), json!({"payload":"start"}))
-        .await
-        .expect("create run");
-
-    let first = manager
-        .send_actor_message(SendActorMessageInput {
-            run_id: &run.id,
-            from_actor_id: "planner",
-            from_peer_id: ACTOR_MAIN_PEER_ID,
-            to_actor_id: "reviewer",
-            to_peer_id: ACTOR_MAIN_PEER_ID,
-            channel: "coordination",
-            transport: TeamActorMessageTransport::Local,
-            route: None,
-            payload: json!({"type":"chat_message","text":"first"}),
-            idempotency_key: None,
-            message_kind: None,
-        })
-        .await
-        .expect("send first message");
-    let second = manager
-        .send_actor_message(SendActorMessageInput {
-            run_id: &run.id,
-            from_actor_id: "planner",
-            from_peer_id: ACTOR_MAIN_PEER_ID,
-            to_actor_id: "reviewer",
-            to_peer_id: ACTOR_MAIN_PEER_ID,
-            channel: "coordination",
-            transport: TeamActorMessageTransport::Local,
-            route: None,
-            payload: json!({"type":"chat_message","text":"second"}),
-            idempotency_key: None,
-            message_kind: None,
-        })
-        .await
-        .expect("send second message");
-    let _other_type = manager
-        .send_actor_message(SendActorMessageInput {
-            run_id: &run.id,
-            from_actor_id: "planner",
-            from_peer_id: ACTOR_MAIN_PEER_ID,
-            to_actor_id: "reviewer",
-            to_peer_id: ACTOR_MAIN_PEER_ID,
-            channel: "coordination",
-            transport: TeamActorMessageTransport::Local,
-            route: None,
-            payload: json!({"type":"  worker_status  ","status":"done"}),
-            idempotency_key: None,
-            message_kind: None,
-        })
-        .await
-        .expect("send other type message");
-
-    let has_other_chat_pending = manager
-        .has_pending_actor_message_payload_type(
-            &run.id,
-            "reviewer",
-            "chat_message",
-            Some(second.message_id),
-        )
-        .await
-        .expect("check chat pending excluding latest");
-    assert!(has_other_chat_pending);
-
-    let has_prior_for_first_chat = manager
-        .has_pending_actor_message_payload_type(
-            &run.id,
-            "reviewer",
-            "chat_message",
-            Some(first.message_id),
-        )
-        .await
-        .expect("check chat pending before first");
-    assert!(!has_prior_for_first_chat);
-
-    manager
-        .ack_actor_message(&run.id, "reviewer", first.message_id)
-        .await
-        .expect("ack first");
-    let still_has_other_chat_pending = manager
-        .has_pending_actor_message_payload_type(
-            &run.id,
-            "reviewer",
-            "chat_message",
-            Some(second.message_id),
-        )
-        .await
-        .expect("check chat pending after ack");
-    assert!(!still_has_other_chat_pending);
-
-    let has_worker_status_pending = manager
-        .has_pending_actor_message_payload_type(&run.id, "reviewer", "worker_status", None)
-        .await
-        .expect("check worker_status pending");
-    assert!(has_worker_status_pending);
-}
-
-#[tokio::test]
-async fn actor_ack_reports_noop_when_message_is_already_delivered() {
-    let db = setup_test_db().await;
-    let manager = TeamManager::new(db);
-
-    let team = manager
-        .create_team(TeamDefinitionConfig {
-            name: "actor-ack-noop-team".to_string(),
-            description: Some("team for duplicate ack diagnostics".to_string()),
-            spec: json!({
-                "entrypoint":"planner",
-                "members":[{"member_id":"planner"},{"member_id":"reviewer"}]
-            }),
-        })
-        .await
-        .expect("create team");
-    let run = manager
-        .create_run(&team.id, Some("ctx-ack-noop"), json!({"payload":"start"}))
-        .await
-        .expect("create run");
-    let sent = manager
-        .send_actor_message(SendActorMessageInput {
-            run_id: &run.id,
-            from_actor_id: "planner",
-            from_peer_id: ACTOR_MAIN_PEER_ID,
-            to_actor_id: "reviewer",
-            to_peer_id: ACTOR_MAIN_PEER_ID,
-            channel: "coordination",
-            transport: TeamActorMessageTransport::Local,
-            route: None,
-            payload: json!({"type":"chat_message","text":"please review"}),
-            idempotency_key: None,
-            message_kind: None,
-        })
-        .await
-        .expect("send message");
-
-    let first = manager
-        .ack_actor_message(&run.id, "reviewer", sent.message_id)
-        .await
-        .expect("first ack");
-    assert!(first.status_changed);
-    assert_eq!(first.message.status, TeamActorMessageStatus::Delivered);
-
-    let second = manager
-        .ack_actor_message(&run.id, "reviewer", sent.message_id)
-        .await
-        .expect("second ack");
-    assert!(!second.status_changed);
-    assert_eq!(second.message.status, TeamActorMessageStatus::Delivered);
-}
-
-#[tokio::test]
-async fn actor_mailbox_service_returns_contract_responses() {
-    let db = setup_test_db().await;
-    let manager = TeamManager::new(db.clone());
-    let service = manager.actor_mailbox_service();
-
-    let team = manager
-        .create_team(TeamDefinitionConfig {
-            name: "actor-mailbox-service-team".to_string(),
-            description: Some("team for actor mailbox service contract".to_string()),
-            spec: json!({
-                "entrypoint":"planner",
-                "members":[{"member_id":"planner"},{"member_id":"reviewer"}]
-            }),
-        })
-        .await
-        .expect("create team");
-    let run = manager
-        .create_run(
-            &team.id,
-            Some("ctx-msg-service"),
-            json!({"payload":"start"}),
-        )
-        .await
-        .expect("create run");
-
-    let sent = service
-        .actor_send(ActorSendRequest {
-            run_id: run.id.clone(),
-            from_actor_id: "planner".to_string(),
-            from_peer_id: None,
-            to_actor_id: Some("reviewer".to_string()),
-            channel_id: None,
-            to_peer_id: None,
-            channel: Some("coordination".to_string()),
-            transport: Some(TeamActorMessageTransport::Local),
-            route: None,
-            payload: json!({"text":"please review"}),
-            idempotency_key: Some("msg-service-1".to_string()),
-            message_kind: None,
-        })
-        .await
-        .expect("actor send");
-    assert_eq!(sent.state, TeamActorMessageStatus::Pending);
-    assert!(!sent.deduped);
-
-    let deduped = service
-        .actor_send(ActorSendRequest {
-            run_id: run.id.clone(),
-            from_actor_id: "planner".to_string(),
-            from_peer_id: None,
-            to_actor_id: Some("reviewer".to_string()),
-            channel_id: None,
-            to_peer_id: None,
-            channel: Some("coordination".to_string()),
-            transport: Some(TeamActorMessageTransport::Local),
-            route: None,
-            payload: json!({"text":"please review"}),
-            idempotency_key: Some("msg-service-1".to_string()),
-            message_kind: None,
-        })
-        .await
-        .expect("actor send deduped");
-    assert_eq!(sent.message_id, deduped.message_id);
-    assert!(deduped.deduped);
-
-    let inbox = service
-        .actor_inbox(ActorInboxRequest {
-            run_id: run.id.clone(),
-            actor_id: "reviewer".to_string(),
-            cursor: None,
-            limit: Some(50),
-            states: None,
-        })
-        .await
-        .expect("actor inbox");
-    assert_eq!(inbox.messages.len(), 1);
-    assert_eq!(inbox.messages[0].message_id, sent.message_id);
-    assert_eq!(inbox.next_cursor, Some(sent.message_id));
-    assert_eq!(inbox.pending_count, 1);
-
-    let acked = service
-        .actor_ack(ActorAckRequest {
-            run_id: run.id.clone(),
-            actor_id: "reviewer".to_string(),
-            message_id: sent.message_id,
-            ack_token: None,
-            result: None,
-        })
-        .await
-        .expect("actor ack");
-    assert_eq!(acked.message_id, sent.message_id);
-    assert_eq!(acked.state, TeamActorMessageStatus::Delivered);
-}
-
-#[tokio::test]
-async fn actor_mailbox_service_cursor_can_hide_page_messages_without_resetting_pending_count() {
-    let db = setup_test_db().await;
-    let manager = TeamManager::new(db.clone());
-    let service = manager.actor_mailbox_service();
-
-    let team = manager
-        .create_team(TeamDefinitionConfig {
-            name: "actor-mailbox-cursor-team".to_string(),
-            description: Some("team for actor inbox cursor semantics".to_string()),
-            spec: json!({
-                "entrypoint":"planner",
-                "members":[{"member_id":"planner"},{"member_id":"reviewer"}]
-            }),
-        })
-        .await
-        .expect("create team");
-    let run = manager
-        .create_run(&team.id, Some("ctx-msg-cursor"), json!({"payload":"start"}))
-        .await
-        .expect("create run");
-
-    let sent = service
-        .actor_send(ActorSendRequest {
-            run_id: run.id.clone(),
-            from_actor_id: "planner".to_string(),
-            from_peer_id: None,
-            to_actor_id: Some("reviewer".to_string()),
-            channel_id: None,
-            to_peer_id: None,
-            channel: Some("coordination".to_string()),
-            transport: Some(TeamActorMessageTransport::Local),
-            route: None,
-            payload: json!({"text":"please review"}),
-            idempotency_key: Some("msg-cursor-1".to_string()),
-            message_kind: None,
-        })
-        .await
-        .expect("actor send");
-
-    let inbox = service
-        .actor_inbox(ActorInboxRequest {
-            run_id: run.id.clone(),
-            actor_id: "reviewer".to_string(),
-            cursor: Some(sent.message_id),
-            limit: Some(50),
-            states: None,
-        })
-        .await
-        .expect("actor inbox with cursor");
-    assert!(inbox.messages.is_empty());
-    assert_eq!(inbox.next_cursor, None);
-    assert_eq!(inbox.pending_count, 1);
-}
-
-#[tokio::test]
-async fn actor_mailbox_service_include_delivered_keeps_pending_visible_on_first_page() {
-    let db = setup_test_db().await;
-    let manager = TeamManager::new(db.clone());
-    let service = manager.actor_mailbox_service();
-
-    let team = manager
-        .create_team(TeamDefinitionConfig {
-            name: "actor-mailbox-pending-first-team".to_string(),
-            description: Some("team for delivered inbox pending-first behavior".to_string()),
-            spec: json!({
-                "entrypoint":"planner",
-                "members":[{"member_id":"planner"},{"member_id":"reviewer"}]
-            }),
-        })
-        .await
-        .expect("create team");
-    let run = manager
-        .create_run(
-            &team.id,
-            Some("ctx-msg-pending-first"),
-            json!({"payload":"start"}),
-        )
-        .await
-        .expect("create run");
-
-    let mut latest_pending_id = None;
-    for idx in 0..25 {
-        let sent = service
-            .actor_send(ActorSendRequest {
-                run_id: run.id.clone(),
-                from_actor_id: "planner".to_string(),
-                from_peer_id: None,
-                to_actor_id: Some("reviewer".to_string()),
-                channel_id: None,
-                to_peer_id: None,
-                channel: Some("coordination".to_string()),
-                transport: Some(TeamActorMessageTransport::Local),
-                route: None,
-                payload: json!({"text": format!("message-{idx}")}),
-                idempotency_key: Some(format!("msg-pending-first-{idx}")),
-                message_kind: None,
-            })
-            .await
-            .expect("actor send");
-        latest_pending_id = Some(sent.message_id);
-        if idx < 24 {
-            service
-                .actor_ack(ActorAckRequest {
-                    run_id: run.id.clone(),
-                    actor_id: "reviewer".to_string(),
-                    message_id: sent.message_id,
-                    ack_token: None,
-                    result: None,
-                })
-                .await
-                .expect("ack historical message");
-        }
-    }
-
-    let inbox = service
-        .actor_inbox(ActorInboxRequest {
-            run_id: run.id,
-            actor_id: "reviewer".to_string(),
-            cursor: None,
-            limit: Some(20),
-            states: Some(vec![
-                TeamActorMessageStatus::Pending,
-                TeamActorMessageStatus::Delivered,
-            ]),
-        })
-        .await
-        .expect("actor inbox with delivered keeps unread visible");
-
-    assert_eq!(inbox.pending_count, 1);
-    assert_eq!(inbox.messages.len(), 1);
-    assert_eq!(inbox.messages[0].status, TeamActorMessageStatus::Pending);
-    assert_eq!(inbox.messages[0].message_id, latest_pending_id.unwrap());
-}
-
-#[tokio::test]
-async fn actor_mailbox_service_include_delivered_returns_history_when_unread_is_empty() {
-    let db = setup_test_db().await;
-    let manager = TeamManager::new(db.clone());
-    let service = manager.actor_mailbox_service();
-
-    let team = manager
-        .create_team(TeamDefinitionConfig {
-            name: "actor-mailbox-history-only-team".to_string(),
-            description: Some("team for delivered-only inbox history".to_string()),
-            spec: json!({
-                "entrypoint":"planner",
-                "members":[{"member_id":"planner"},{"member_id":"reviewer"}]
-            }),
-        })
-        .await
-        .expect("create team");
-    let run = manager
-        .create_run(
-            &team.id,
-            Some("ctx-msg-history-only"),
-            json!({"payload":"start"}),
-        )
-        .await
-        .expect("create run");
-
-    let mut first_delivered_id = None;
-    for idx in 0..3 {
-        let sent = service
-            .actor_send(ActorSendRequest {
-                run_id: run.id.clone(),
-                from_actor_id: "planner".to_string(),
-                from_peer_id: None,
-                to_actor_id: Some("reviewer".to_string()),
-                channel_id: None,
-                to_peer_id: None,
-                channel: Some("coordination".to_string()),
-                transport: Some(TeamActorMessageTransport::Local),
-                route: None,
-                payload: json!({"text": format!("history-{idx}")}),
-                idempotency_key: Some(format!("msg-history-only-{idx}")),
-                message_kind: None,
-            })
-            .await
-            .expect("actor send");
-        if first_delivered_id.is_none() {
-            first_delivered_id = Some(sent.message_id);
-        }
-        service
-            .actor_ack(ActorAckRequest {
-                run_id: run.id.clone(),
-                actor_id: "reviewer".to_string(),
-                message_id: sent.message_id,
-                ack_token: None,
-                result: None,
-            })
-            .await
-            .expect("ack historical message");
-    }
-
-    let inbox = service
-        .actor_inbox(ActorInboxRequest {
-            run_id: run.id,
-            actor_id: "reviewer".to_string(),
-            cursor: None,
-            limit: Some(20),
-            states: Some(vec![
-                TeamActorMessageStatus::Pending,
-                TeamActorMessageStatus::Delivered,
-            ]),
-        })
-        .await
-        .expect("actor inbox with delivered history only");
-
-    assert_eq!(inbox.pending_count, 0);
-    assert_eq!(inbox.messages.len(), 3);
-    assert!(
-        inbox
-            .messages
-            .iter()
-            .all(|message| message.status == TeamActorMessageStatus::Delivered)
-    );
-    assert_eq!(inbox.messages[0].message_id, first_delivered_id.unwrap());
-}
-
-#[tokio::test]
-async fn actor_mailbox_service_include_delivered_preserves_requested_mix_when_page_has_pending() {
-    let db = setup_test_db().await;
-    let manager = TeamManager::new(db.clone());
-    let service = manager.actor_mailbox_service();
-
-    let team = manager
-        .create_team(TeamDefinitionConfig {
-            name: "actor-mailbox-mixed-first-page-team".to_string(),
-            description: Some("team for delivered inbox mixed first page".to_string()),
-            spec: json!({
-                "entrypoint":"planner",
-                "members":[{"member_id":"planner"},{"member_id":"reviewer"}]
-            }),
-        })
-        .await
-        .expect("create team");
-    let run = manager
-        .create_run(
-            &team.id,
-            Some("ctx-msg-mixed-first-page"),
-            json!({"payload":"start"}),
-        )
-        .await
-        .expect("create run");
-
-    for idx in 0..3 {
-        let sent = service
-            .actor_send(ActorSendRequest {
-                run_id: run.id.clone(),
-                from_actor_id: "planner".to_string(),
-                from_peer_id: None,
-                to_actor_id: Some("reviewer".to_string()),
-                channel_id: None,
-                to_peer_id: None,
-                channel: Some("coordination".to_string()),
-                transport: Some(TeamActorMessageTransport::Local),
-                route: None,
-                payload: json!({"text": format!("mixed-{idx}")}),
-                idempotency_key: Some(format!("msg-mixed-first-page-{idx}")),
-                message_kind: None,
-            })
-            .await
-            .expect("actor send");
-        if idx < 2 {
-            service
-                .actor_ack(ActorAckRequest {
-                    run_id: run.id.clone(),
-                    actor_id: "reviewer".to_string(),
-                    message_id: sent.message_id,
-                    ack_token: None,
-                    result: None,
-                })
-                .await
-                .expect("ack historical message");
-        }
-    }
-
-    let inbox = service
-        .actor_inbox(ActorInboxRequest {
-            run_id: run.id,
-            actor_id: "reviewer".to_string(),
-            cursor: None,
-            limit: Some(20),
-            states: Some(vec![
-                TeamActorMessageStatus::Pending,
-                TeamActorMessageStatus::Delivered,
-            ]),
-        })
-        .await
-        .expect("actor inbox with delivered mixed page");
-
-    assert_eq!(inbox.pending_count, 1);
-    assert_eq!(inbox.messages.len(), 3);
-    assert_eq!(inbox.messages[0].status, TeamActorMessageStatus::Delivered);
-    assert_eq!(inbox.messages[1].status, TeamActorMessageStatus::Delivered);
-    assert_eq!(inbox.messages[2].status, TeamActorMessageStatus::Pending);
-}
-
-#[tokio::test]
-async fn actor_mailbox_service_triage_hides_watching_messages_from_unread_snapshot() {
-    let db = setup_test_db().await;
-    let manager = TeamManager::new(db.clone());
-    let service = manager.actor_mailbox_service();
-
-    let team = manager
-        .create_team(TeamDefinitionConfig {
-            name: "actor-mailbox-triage-team".to_string(),
-            description: Some("team for mailbox triage unread semantics".to_string()),
-            spec: json!({
-                "entrypoint":"planner",
-                "members":[{"member_id":"planner"},{"member_id":"reviewer"}]
-            }),
-        })
-        .await
-        .expect("create team");
-    let run = manager
-        .create_run(&team.id, Some("ctx-msg-triage"), json!({"payload":"start"}))
-        .await
-        .expect("create run");
-
-    let sent = service
-        .actor_send(ActorSendRequest {
-            run_id: run.id.clone(),
-            from_actor_id: "planner".to_string(),
-            from_peer_id: None,
-            to_actor_id: Some("reviewer".to_string()),
-            channel_id: None,
-            to_peer_id: None,
-            channel: Some("coordination".to_string()),
-            transport: Some(TeamActorMessageTransport::Local),
-            route: None,
-            payload: json!({"text":"observe this request"}),
-            idempotency_key: Some("msg-triage-watch-1".to_string()),
-            message_kind: None,
-        })
-        .await
-        .expect("actor send");
-
-    let triaged = service
-        .actor_triage(ActorTriageRequest {
-            run_id: run.id.clone(),
-            actor_id: "reviewer".to_string(),
-            message_id: sent.message_id,
-            disposition: ActorMessageHandlingDisposition::Watching,
-        })
-        .await
-        .expect("triage watching");
-    assert_eq!(
-        triaged.disposition,
-        ActorMessageHandlingDisposition::Watching
-    );
-    assert!(triaged.handling_changed);
-    assert_eq!(
-        triaged.message.handled_by_actor_id.as_deref(),
-        Some("reviewer")
-    );
-
-    let unread = service
-        .actor_inbox(ActorInboxRequest {
-            run_id: run.id.clone(),
-            actor_id: "reviewer".to_string(),
-            cursor: None,
-            limit: Some(20),
-            states: None,
-        })
-        .await
-        .expect("unread inbox after watch triage");
-    assert_eq!(unread.pending_count, 0);
-    assert!(unread.messages.is_empty());
-
-    let with_history = service
-        .actor_inbox(ActorInboxRequest {
-            run_id: run.id,
-            actor_id: "reviewer".to_string(),
-            cursor: None,
-            limit: Some(20),
-            states: Some(vec![
-                TeamActorMessageStatus::Pending,
-                TeamActorMessageStatus::Delivered,
-            ]),
-        })
-        .await
-        .expect("history inbox after watch triage");
-    assert_eq!(with_history.messages.len(), 1);
-    assert_eq!(
-        with_history.messages[0].handling_disposition,
-        ActorMessageHandlingDisposition::Watching
-    );
-}
-
-#[tokio::test]
-async fn actor_mailbox_service_claims_topics_and_prevents_parallel_takeover() {
-    let db = setup_test_db().await;
-    let manager = TeamManager::new(db.clone());
-    let service = manager.actor_mailbox_service();
-
-    let team = manager
-        .create_team(TeamDefinitionConfig {
-            name: "actor-mailbox-claim-team".to_string(),
-            description: Some("team for mailbox thread claim semantics".to_string()),
-            spec: json!({
-                "entrypoint":"planner",
-                "members":[
-                    {"member_id":"planner"},
-                    {"member_id":"reviewer"},
-                    {"member_id":"worker"}
-                ]
-            }),
-        })
-        .await
-        .expect("create team");
-    let run = manager
-        .create_run(&team.id, Some("ctx-msg-claim"), json!({"payload":"start"}))
-        .await
-        .expect("create run");
-    let task_title = "Review design".to_string();
-    let (task, _) = manager
-        .create_task_with_metadata(crate::team::TeamTaskCreateInput {
-            team_id: &team.id,
-            title: &task_title,
-            created_by_actor_id: "planner",
-            priority: crate::team::TeamTaskPriority::High,
-            assigned_member_id: Some("reviewer"),
-            context: json!({}),
-            conversation_mode: "group_chat",
-            topic: Some("mailbox claim"),
-        })
-        .await
-        .expect("create task");
-
-    let reviewer_message = service
-        .actor_send(ActorSendRequest {
-            run_id: run.id.clone(),
-            from_actor_id: "planner".to_string(),
-            from_peer_id: None,
-            to_actor_id: Some("reviewer".to_string()),
-            channel_id: None,
-            to_peer_id: None,
-            channel: Some("coordination".to_string()),
-            transport: Some(TeamActorMessageTransport::Local),
-            route: None,
-            payload: json!({
-                "text":"please take point on this topic",
-                "task_id": task.id.clone(),
-                "task_message_id": 77
-            }),
-            idempotency_key: Some("msg-claim-reviewer".to_string()),
-            message_kind: None,
-        })
-        .await
-        .expect("send reviewer message");
-    let worker_message = service
-        .actor_send(ActorSendRequest {
-            run_id: run.id.clone(),
-            from_actor_id: "planner".to_string(),
-            from_peer_id: None,
-            to_actor_id: Some("worker".to_string()),
-            channel_id: None,
-            to_peer_id: None,
-            channel: Some("coordination".to_string()),
-            transport: Some(TeamActorMessageTransport::Local),
-            route: None,
-            payload: json!({
-                "text":"same topic for another observer",
-                "task_id": task.id.clone(),
-                "task_message_id": 77
-            }),
-            idempotency_key: Some("msg-claim-worker".to_string()),
-            message_kind: None,
-        })
-        .await
-        .expect("send worker message");
-
-    let reviewer_claim = service
-        .actor_triage(ActorTriageRequest {
-            run_id: run.id.clone(),
-            actor_id: "reviewer".to_string(),
-            message_id: reviewer_message.message_id,
-            disposition: ActorMessageHandlingDisposition::Claimed,
-        })
-        .await
-        .expect("reviewer claim");
-    assert_eq!(
-        reviewer_claim.message.thread_owner_actor_id.as_deref(),
-        Some("reviewer")
-    );
-    assert_eq!(
-        reviewer_claim.message.thread_claim_status,
-        Some(agenthub_team_actor::ActorThreadClaimStatus::Claimed)
-    );
-
-    let err = service
-        .actor_triage(ActorTriageRequest {
-            run_id: run.id.clone(),
-            actor_id: "worker".to_string(),
-            message_id: worker_message.message_id,
-            disposition: ActorMessageHandlingDisposition::Claimed,
-        })
-        .await
-        .expect_err("parallel takeover should conflict");
-    assert_eq!(err.code, ActorServiceErrorCode::Conflict);
-    assert!(err.message.contains("reviewer"));
-
-    service
-        .actor_triage(ActorTriageRequest {
-            run_id: run.id.clone(),
-            actor_id: "reviewer".to_string(),
-            message_id: reviewer_message.message_id,
-            disposition: ActorMessageHandlingDisposition::Released,
-        })
-        .await
-        .expect("release claim");
-
-    let worker_claim = service
-        .actor_triage(ActorTriageRequest {
-            run_id: run.id,
-            actor_id: "worker".to_string(),
-            message_id: worker_message.message_id,
-            disposition: ActorMessageHandlingDisposition::Claimed,
-        })
-        .await
-        .expect("worker claim after release");
-    assert_eq!(
-        worker_claim.message.thread_owner_actor_id.as_deref(),
-        Some("worker")
-    );
-}
-
-#[tokio::test]
-async fn actor_mailbox_service_requires_active_owner_for_release_and_complete() {
-    let db = setup_test_db().await;
-    let manager = TeamManager::new(db.clone());
-    let service = manager.actor_mailbox_service();
-
-    let team = manager
-        .create_team(TeamDefinitionConfig {
-            name: "actor-mailbox-release-team".to_string(),
-            description: Some("team for mailbox ownership guardrails".to_string()),
-            spec: json!({
-                "entrypoint":"planner",
-                "members":[
-                    {"member_id":"planner"},
-                    {"member_id":"reviewer"},
-                    {"member_id":"worker"}
-                ]
-            }),
-        })
-        .await
-        .expect("create team");
-    let run = manager
-        .create_run(
-            &team.id,
-            Some("ctx-msg-release-guard"),
-            json!({"payload":"start"}),
-        )
-        .await
-        .expect("create run");
-
-    let reviewer_message = service
-        .actor_send(ActorSendRequest {
-            run_id: run.id.clone(),
-            from_actor_id: "planner".to_string(),
-            from_peer_id: None,
-            to_actor_id: Some("reviewer".to_string()),
-            channel_id: None,
-            to_peer_id: None,
-            channel: Some("coordination".to_string()),
-            transport: Some(TeamActorMessageTransport::Local),
-            route: None,
-            payload: json!({
-                "text":"reviewer owns this topic",
-                "correlation_id":"release-guard-1"
-            }),
-            idempotency_key: Some("msg-release-guard-reviewer".to_string()),
-            message_kind: None,
-        })
-        .await
-        .expect("send reviewer message");
-    let worker_message = service
-        .actor_send(ActorSendRequest {
-            run_id: run.id.clone(),
-            from_actor_id: "planner".to_string(),
-            from_peer_id: None,
-            to_actor_id: Some("worker".to_string()),
-            channel_id: None,
-            to_peer_id: None,
-            channel: Some("coordination".to_string()),
-            transport: Some(TeamActorMessageTransport::Local),
-            route: None,
-            payload: json!({
-                "text":"worker sees the same topic",
-                "correlation_id":"release-guard-1"
-            }),
-            idempotency_key: Some("msg-release-guard-worker".to_string()),
-            message_kind: None,
-        })
-        .await
-        .expect("send worker message");
-
-    service
-        .actor_triage(ActorTriageRequest {
-            run_id: run.id.clone(),
-            actor_id: "reviewer".to_string(),
-            message_id: reviewer_message.message_id,
-            disposition: ActorMessageHandlingDisposition::Claimed,
-        })
-        .await
-        .expect("reviewer claim");
-
-    let release_err = service
-        .actor_triage(ActorTriageRequest {
-            run_id: run.id.clone(),
-            actor_id: "worker".to_string(),
-            message_id: worker_message.message_id,
-            disposition: ActorMessageHandlingDisposition::Released,
-        })
-        .await
-        .expect_err("non-owner release should conflict");
-    assert_eq!(release_err.code, ActorServiceErrorCode::Conflict);
-    assert!(release_err.message.contains("reviewer"));
-
-    let complete_err = service
-        .actor_triage(ActorTriageRequest {
-            run_id: run.id.clone(),
-            actor_id: "worker".to_string(),
-            message_id: worker_message.message_id,
-            disposition: ActorMessageHandlingDisposition::Completed,
-        })
-        .await
-        .expect_err("non-owner complete should conflict");
-    assert_eq!(complete_err.code, ActorServiceErrorCode::Conflict);
-    assert!(complete_err.message.contains("reviewer"));
-
-    let worker_history = service
-        .actor_inbox(ActorInboxRequest {
-            run_id: run.id,
-            actor_id: "worker".to_string(),
-            cursor: None,
-            limit: Some(20),
-            states: Some(vec![TeamActorMessageStatus::Pending]),
-        })
-        .await
-        .expect("worker history inbox");
-    assert_eq!(worker_history.messages.len(), 1);
-    assert_eq!(
-        worker_history.messages[0].handling_disposition,
-        ActorMessageHandlingDisposition::Untriaged
-    );
-    assert_eq!(
-        worker_history.messages[0].thread_owner_actor_id.as_deref(),
-        Some("reviewer")
-    );
-}
-
-#[tokio::test]
-async fn actor_mailbox_service_completed_claim_remains_visible_in_history() {
-    let db = setup_test_db().await;
-    let manager = TeamManager::new(db.clone());
-    let service = manager.actor_mailbox_service();
-
-    let team = manager
-        .create_team(TeamDefinitionConfig {
-            name: "actor-mailbox-completed-team".to_string(),
-            description: Some("team for completed topic visibility".to_string()),
-            spec: json!({
-                "entrypoint":"planner",
-                "members":[{"member_id":"planner"},{"member_id":"reviewer"}]
-            }),
-        })
-        .await
-        .expect("create team");
-    let run = manager
-        .create_run(
-            &team.id,
-            Some("ctx-msg-completed-visible"),
-            json!({"payload":"start"}),
-        )
-        .await
-        .expect("create run");
-
-    let sent = service
-        .actor_send(ActorSendRequest {
-            run_id: run.id.clone(),
-            from_actor_id: "planner".to_string(),
-            from_peer_id: None,
-            to_actor_id: Some("reviewer".to_string()),
-            channel_id: None,
-            to_peer_id: None,
-            channel: Some("coordination".to_string()),
-            transport: Some(TeamActorMessageTransport::Local),
-            route: None,
-            payload: json!({
-                "text":"finish this mailbox topic",
-                "correlation_id":"completed-visible-1"
-            }),
-            idempotency_key: Some("msg-completed-visible-reviewer".to_string()),
-            message_kind: None,
-        })
-        .await
-        .expect("send mailbox message");
-
-    service
-        .actor_triage(ActorTriageRequest {
-            run_id: run.id.clone(),
-            actor_id: "reviewer".to_string(),
-            message_id: sent.message_id,
-            disposition: ActorMessageHandlingDisposition::Claimed,
-        })
-        .await
-        .expect("claim mailbox topic");
-    let completed = service
-        .actor_triage(ActorTriageRequest {
-            run_id: run.id.clone(),
-            actor_id: "reviewer".to_string(),
-            message_id: sent.message_id,
-            disposition: ActorMessageHandlingDisposition::Completed,
-        })
-        .await
-        .expect("complete mailbox topic");
-    assert_eq!(
-        completed.message.handling_disposition,
-        ActorMessageHandlingDisposition::Completed
-    );
-
-    let unread = service
-        .actor_inbox(ActorInboxRequest {
-            run_id: run.id.clone(),
-            actor_id: "reviewer".to_string(),
-            cursor: None,
-            limit: Some(20),
-            states: None,
-        })
-        .await
-        .expect("unread inbox after completed triage");
-    assert_eq!(unread.pending_count, 0);
-    assert!(unread.messages.is_empty());
-
-    let history = service
-        .actor_inbox(ActorInboxRequest {
-            run_id: run.id,
-            actor_id: "reviewer".to_string(),
-            cursor: None,
-            limit: Some(20),
-            states: Some(vec![
-                TeamActorMessageStatus::Pending,
-                TeamActorMessageStatus::Delivered,
-            ]),
-        })
-        .await
-        .expect("history inbox after completed triage");
-    assert_eq!(history.messages.len(), 1);
-    assert_eq!(
-        history.messages[0].thread_claim_status,
-        Some(agenthub_team_actor::ActorThreadClaimStatus::Completed)
-    );
-    assert_eq!(
-        history.messages[0].thread_owner_actor_id.as_deref(),
-        Some("reviewer")
-    );
-    assert_eq!(
-        history.messages[0].handling_disposition,
-        ActorMessageHandlingDisposition::Completed
-    );
-}
-
-#[tokio::test]
-async fn actor_mailbox_service_task_link_surfaces_durable_task_association() {
-    let db = setup_test_db().await;
-    let manager = TeamManager::new(db.clone());
-    let service = manager.actor_mailbox_service();
-
-    let team = manager
-        .create_team(TeamDefinitionConfig {
-            name: "actor-mailbox-task-link-team".to_string(),
-            description: Some("team for mailbox task link semantics".to_string()),
-            spec: json!({
-                "entrypoint":"planner",
-                "members":[{"member_id":"planner"},{"member_id":"reviewer"}]
-            }),
-        })
-        .await
-        .expect("create team");
-    let run = manager
-        .create_run(
-            &team.id,
-            Some("ctx-msg-task-link"),
-            json!({"payload":"start"}),
-        )
-        .await
-        .expect("create run");
-    let task_title = "Investigate mailbox".to_string();
-    let (task, _) = manager
-        .create_task_with_metadata(crate::team::TeamTaskCreateInput {
-            team_id: &team.id,
-            title: &task_title,
-            created_by_actor_id: "planner",
-            priority: crate::team::TeamTaskPriority::Medium,
-            assigned_member_id: Some("reviewer"),
-            context: json!({}),
-            conversation_mode: "group_chat",
-            topic: Some("task link"),
-        })
-        .await
-        .expect("create task");
-
-    let sent = service
-        .actor_send(ActorSendRequest {
-            run_id: run.id.clone(),
-            from_actor_id: "planner".to_string(),
-            from_peer_id: None,
-            to_actor_id: Some("reviewer".to_string()),
-            channel_id: None,
-            to_peer_id: None,
-            channel: Some("coordination".to_string()),
-            transport: Some(TeamActorMessageTransport::Local),
-            route: None,
-            payload: json!({"text":"convert this into a tracked lane"}),
-            idempotency_key: Some("msg-task-link-reviewer".to_string()),
-            message_kind: None,
-        })
-        .await
-        .expect("send mailbox message");
-
-    let linked = service
-        .actor_task_link(ActorTaskLinkRequest {
-            run_id: run.id.clone(),
-            actor_id: "reviewer".to_string(),
-            message_id: sent.message_id,
-            task_id: task.id.clone(),
-            relation: ActorMessageTaskRelation::SpawnedTask,
-        })
-        .await
-        .expect("link mailbox message to task");
-    assert!(linked.created);
-    assert_eq!(linked.task_id, task.id);
-    assert_eq!(linked.relation, ActorMessageTaskRelation::SpawnedTask);
-    assert_eq!(
-        linked.message.linked_task_id.as_deref(),
-        Some(task.id.as_str())
-    );
-    assert_eq!(
-        linked.message.linked_task_relation,
-        Some(ActorMessageTaskRelation::SpawnedTask)
-    );
-
-    let link_count: i64 = sqlx::query_scalar(
-        "SELECT COUNT(*) FROM team_actor_message_links WHERE run_id = ?1 AND message_id = ?2 AND task_id = ?3",
-    )
-    .bind(&run.id)
-    .bind(sent.message_id)
-    .bind(&task.id)
-    .fetch_one(&db)
-    .await
-    .expect("count mailbox task links");
-    assert_eq!(link_count, 1);
 }
 
 #[tokio::test]
